presser

Function copy_from_iter_to_offset_with_align_packed

pub fn copy_from_iter_to_offset_with_align_packed<T: Copy, Iter: Iterator<Item = T>, S: Slab>(
    src: Iter,
    dst: &mut S,
    start_offset: usize,
    min_alignment: usize,
) -> Result<Option<CopyRecord>, CopyError>

Like [copy_from_iter_to_offset_with_align] except that alignment between elements yielded by the iterator will ignore min_alignment and rather only be aligned to the alignment of T.

Because of this, only one CopyRecord is returned specifying the record of the entire block of copied data. If the src iterator is empty, returns None.